feat(optimizer): Annotate DAYOFWEEK(expr), DAYOFMONTH(expr) for Hive/Spark/DBX (#6996)

+++ packages/sqlingo.js/tests/upstream/fixtures/optimizer/annotate_functions.sql
@@ -92,7 +92,7 @@ TO_BASE64(tbl.bytes_col);
 VARCHAR;
 
 UNIX_DATE(tbl.date_col);
-BIGINT;
+INT;
 
 UNIX_SECONDS(tbl.timestamp_col);
 BIGINT;
